@@ -4,18 +4,81 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.2",
-        "ext-ctype": "*",
+        "php": "^8.4",
         "ext-iconv": "*",
-        "symfony/console": "7.2.*",
-        "symfony/dotenv": "7.2.*",
-        "symfony/flex": "^2",
-        "symfony/framework-bundle": "7.2.*",
-        "symfony/runtime": "7.2.*",
-        "symfony/yaml": "7.2.*"
+        "api-platform/core": "^4.0.12",
+        "composer/package-versions-deprecated": "^1.11.99.5",
+        "doctrine/annotations": "^2.0.2",
+        "doctrine/doctrine-bundle": "^2.13.1",
+        "doctrine/doctrine-migrations-bundle": "^3.3.1",
+        "doctrine/mongodb-odm-bundle": "^5.1",
+        "doctrine/orm": "^3.3",
+        "kreait/firebase-bundle": ">=5.5",
+        "lexik/jwt-authentication-bundle": "^3.1",
+        "nelmio/cors-bundle": "^2.5",
+        "phpdocumentor/reflection-docblock": "^5.6.1",
+        "phpstan/phpdoc-parser": "^2.0",
+        "symfony/asset": "^7.2",
+        "symfony/console": "^7.2.1",
+        "symfony/debug-bundle": "^7.2",
+        "symfony/dotenv": "^7.2",
+        "symfony/expression-language": "^7.2",
+        "symfony/filesystem": "^7.2",
+        "symfony/finder": "^7.2",
+        "symfony/flex": "^2.4.7",
+        "symfony/form": "^7.2",
+        "symfony/framework-bundle": "^7.2.1",
+        "symfony/intl": "^7.2",
+        "symfony/monolog-bundle": "^3.10",
+        "symfony/process": "^7.2",
+        "symfony/property-access": "^7.2",
+        "symfony/property-info": "^7.2.1",
+        "symfony/runtime": "^7.2",
+        "symfony/security-bundle": "7.2.*",
+        "symfony/serializer": "^7.2",
+        "symfony/stopwatch": "^7.2",
+        "symfony/string": "^7.2",
+        "symfony/uid": "^7.2",
+        "symfony/validator": "^7.2",
+        "symfony/var-dumper": "^7.2",
+        "symfony/web-link": "^7.2",
+        "symfony/web-profiler-bundle": "^7.2",
+        "symfony/yaml": "^7.2",
+        "tegbessou/empire-des-amis-security-authenticator": "^1.1.2",
+        "tegbessou/shared-kernel": "^1.7.2",
+        "twig/extra-bundle": "^3.17",
+        "twig/twig": "^3.17.1"
     },
     "require-dev": {
+        "friends-of-behat/page-object-extension": "^0.3.2",
+        "friends-of-behat/symfony-extension": "^2.6",
+        "friendsofphp/php-cs-fixer": "^3.65",
+        "phpmd/phpmd": "^2.15",
+        "phpstan/phpstan": "^2.0.4",
+        "phpstan/phpstan-dibi": "^2.0",
+        "phpstan/phpstan-doctrine": "^2.0.1",
+        "phpstan/phpstan-phpunit": "^2.0.2",
+        "phpstan/phpstan-strict-rules": "^2.0.1",
+        "phpunit/phpunit": "^11.5.1",
+        "qossmic/deptrac": "^2.0.4",
+        "rector/rector": "^2.0.0-rc2",
+        "symfony/browser-kit": "7.2.*",
+        "symfony/css-selector": "^7.2",
+        "symfony/http-client": "7.2.*",
+        "symfony/maker-bundle": "^1.61",
+        "symfony/phpunit-bridge": "^7.2",
+        "zenstruck/messenger-test": "^1.11"
     },
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "https://github.com/tegbessou/shared-kernel.git"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/tegbessou/empire-des-amis-security-authenticator.git"
+        }
+    ],
     "config": {
         "allow-plugins": {
             "php-http/discovery": true,
@@ -27,12 +90,13 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "EmpireDesAmis\\Country\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "App\\Tests\\": "tests/"
+            "EmpireDesAmis\\Country\\Tests\\": "tests/",
+            "EmpireDesAmis\\Country\\DataFixtures\\": "fixtures/"
         }
     },
     "replace": {
